Delete authorization, listings ordering

+++ App/controllers/ListingController.php
@@ -1,10 +1,12 @@
 <?php
- declare(strict_types=1);
+  declare(strict_types=1);
 
 namespace App\Controllers;
 
 use Framework\Database;
 use Framework\Validation;
+use Framework\Session;
+use Framework\Authorization;
 
 class ListingController
 {
@@ -24,7 +26,7 @@ class ListingController
     public function index()
     {
 
-        $listings = $this->db->query('SELECT * FROM listings')->fetchAll();
+        $listings = $this->db->query('SELECT * FROM listings ORDER BY created_at DESC')->fetchAll();
 
         loadView('listings/index', ['listings' => $listings]);
     }
@@ -77,7 +79,7 @@ class ListingController
 
         $newListingData = array_intersect_key($_POST, array_flip($allowedFields));
 
-        $newListingData['user_id'] = 1;
+        $newListingData['user_id'] = Session::get('user')['id'];
 
         $newListingData = array_map('sanitize', $newListingData);
 
@@ -137,9 +139,16 @@ class ListingController
 
         $listing = $this->db->query('SELECT * FROM listings WHERE id = :id', $params)->fetch();
 
+        // Check if the listing exists
         if (!$listing) {
             ErrorController::notFound('Listing not found');
             return;
+        }
+
+        // Authorization
+        if(!Authorization::isOwner($listing->user_id)) {
+            $_SESSION['error_message'] = 'You are not allowed to delete this listing';
+            return redirect('/listings/' .$listing->id);
         }
 
         $this->db->query('DELETE FROM listings WHERE id = :id', $params);
@@ -149,8 +158,6 @@ class ListingController
 
         redirect('/listings');
     }
-
-
 
     /**
      * Show the listing edit form
